moved getThemeFromLocalStore to utils.js
