+++ src/components/AppContext.jsx
@@ -1,22 +1,8 @@
 
 import React, { createContext, useContext, useState } from 'react';
+import { getThemeFromLocalStorage } from '../utils/utils';
 
 const AppContext = createContext();
-
-// TODO: move to utils
-// get the theme from local storage
-//      is used to keep the theme set on page refresh
-function getThemeFromLocalStorage() {
-    let theme = localStorage.getItem('theme');
-
-    if (theme == null || theme === "light-theme") {
-        theme = "light-theme";
-        document.body.classList.remove("dark-theme");
-    } else {
-        document.body.classList.add("dark-theme");
-    }
-    return theme;
-}
 
 /*  ===============================================
  *  COMPONENT DEFINITION
